hallelujah! use .tar.gz on windows too
